requirements.txt
streamlit
paddleocr
ultralytics
pillow
pdf2image
pymupdf
numpy
paddlepaddle
fitz
torch=2.3.1
# python=3.9